Make FormattedMethodMiner extends Executor to control its resources
