+++ kotlin-analysis-clones/src/main/kotlin/org/jetbrains/research/ml/kotlinAnalysis/IndexBuilder.kt
@@ -3,7 +3,6 @@ package org.jetbrains.research.ml.kotlinAnalysis
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiDocumentManager
 import org.jetbrains.kotlin.psi.KtNamedFunction
-import org.jetbrains.research.ml.kotlinAnalysis.util.getPrintWriter
 import java.nio.file.Path
 import java.nio.file.Paths
 
@@ -11,14 +10,14 @@ import java.nio.file.Paths
  * Builds incremental index of given project or method.
  * Index is printed to file in output directory.
  */
-class IndexBuilder(outputDir: Path) : AutoCloseable {
+class IndexBuilder(outputDir: Path) : Resource {
     private var lastProjectId = 0
     private var lastMethodId = 0
-    private val projectIndexWriter = getPrintWriter(outputDir, "project_index.csv")
-    private val methodIndexWriter = getPrintWriter(outputDir, "method_index.csv")
+    private val projectIndexWriter = PrintWriterRecourse(outputDir, "project_index.csv")
+    private val methodIndexWriter = PrintWriterRecourse(outputDir, "method_index.csv")
 
     fun indexProject(project: Project): Int {
-        projectIndexWriter.println("$lastProjectId\t${project.name}")
+        projectIndexWriter.writer.println("$lastProjectId\t${project.name}")
         return lastProjectId++
     }
 
@@ -33,10 +32,15 @@ class IndexBuilder(outputDir: Path) : AutoCloseable {
             ?: throw IllegalArgumentException("Cannot find document containing function ${function.name}")
         val startLine = doc.getLineNumber(function.textRange.startOffset)
         val endLine = doc.getLineNumber(function.textRange.endOffset)
-        methodIndexWriter.println(
+        methodIndexWriter.writer.println(
             "$projectId\t$lastMethodId\t$fileRelativePath\t$startLine\t$endLine"
         )
         return lastMethodId++
+    }
+
+    override fun init() {
+        projectIndexWriter.init()
+        methodIndexWriter.init()
     }
 
     override fun close() {
